Before: Don't account for a hit of the cerenkov slab when it is a neutral particle and the energy is < 100keV. This definition is targeted to filter out optical photons.
Now: Don't account for a hit of the cerenkov slab when it is an optical photon.
 
This hit filter is needed for a correct hit counting, otherwise each internal reflection of a cerenkov photon crosses boundary and will be accounted an a valid hit. The current definition allows now also slow neutrons to be accounted correctly (in case we need it, i.e. neutrons causing scintillation light).