In order to visualize the experiment better for the open house on May 1, 2010 I hid the collimator support walls.
I also added a stripped down version of the visualization macro.  With this macro the trajectories are not drawn over the detector volumes.  Instead the track is in its proper place.  I found this edit to be the most important for the open house because although it hides some of the tracks behind collimator walls and magnets it is more physically interpretable.