Modification to the magnetic field routines.  This potentially affects macros.
The magnetic field maps can be read in from the ANSYS dat files, but now also
from the binary field map files that the tracking analyzer uses (faster).  The
underlying structures used for the field map interpolation have been adopted
from the tracking analyzer as well.

The existing messenger class for the magnetic field map was rewritten and
activated.  The field map MainMagnet_FieldMap.dat is now not loaded auto-
matically at the start, but only after you explicitly issue a /MagneticField/Read
command (with optional field map file name).  The command /MagneticField/SetScale
can be used to rescale the field.

This now allows for the following macro snippet:

  # Trigger on Cerenkov events
  /Trigger/Enable cer
  # Load binary field map
  /MagneticField/Read MainMagnet_FieldMap.bin
  # Set output filename stem
  /Analysis/RootFileStem QtorScan_
  # file: QtorScan_1.root
  /MagneticField/SetScale 0.8
  /run/beamOn 10000
  # file: QtorScan_2.root
  /MagneticField/SetScale 0.9
  /run/beamOn 10000
  # file: QtorScan_3.root
  /MagneticField/SetScale 1.0
  /run/beamOn 10000
  # file: QtorScan_4.root
  /MagneticField/SetScale 1.1
  /run/beamOn 10000

The macro myQweakConfiguration.mac was updated to load the usual field map file
MainMagnet_FieldMap.dat.

The magnetic field messenger should allow to set different field map grid sizes
and rotations more easily (but that isn't implemented here yet).  That is what
would be needed to get this to read the 2010 and 2011 field maps as well.