Added support for all field maps currently in use.
To load a particular field map, use for example
  /control/execute macros/peiqing_2011.mac
This will look for the file peiqing_2011.dat in the current directory,
so be sure to place a link there.

The necessary work was adding messenger functions to set the field map
grid parameters.  These are all in /MagneticField/...

The field map macros in macros/ also set the reference current for that
field map, and /MagneticField/SetActualCurrent 8921 can be used to set
a particular actual current.

There is good agreement between the fields loaded by macros/peiqing_2007.mac
and macros/peiqing_2011.mac, but the 2010 fields seem different by about
0.1 kG.  Maybe I have the reference currents wrong or so.  Since 2007 and
2011 are preferred anyway, this shouldn't matter so much.