Many modifications were made in this commit, mainly for fixing bugs in the generator and preventing potential issues associated with the operations in the user stepping action.
Now the primary events are generated in two seperated steps. In the first step, an electron with 1.165 GeV beam energy is fired from upstream of target. When it reaches a random generated z-location in the target (this z-location will be the z-location of scattering vertex), its energy loss (E_loss) is recorded. Then the scattering information (e.g. scattering angle and cross-section etc.) is computed, and the track is stopped and killed. In the second step, another electron is fired from the scattering vertex with an energy of (1.165 - E_loss) GeV. This electron and its secondaries will be tracked. 

This method throw out half of total primaries to simulate the energy loss in the target. Because we do not trace these events, they only occupy a very small portion of computing time.

Other modifications in this commit include:
- fixed the theta/phi angle calculations in the detector branches.
- turned opticalphoton process on.
- default target center position is set to z=-652.67 cm.