The default version of Geant4 from 12gev_phys is now 4.9.5.  On ifarml64 it is a non-working 4.9.5 (maintainer knows about it, ticket with CC is in).  Geant 4.9.5 throws the master makefile method out of the window in favor of cmake.  This is a first attempt at a CMakeLists.txt file (the equivalent of the GNUmakefile).
*** RECOMMENDED ***
To revert to an older version of Geant4, use
  setenv GEANT4_VERSION 4.9.4.p02
  setenv JLAB_ROOT /site/12gev_phys
  source ${JLAB_ROOT}/ce/jlab.csh
  cd <wherever the source lives>
  make


On jlabl1 (but not ifarml64):
  setenv GEANT4_VERSION 4.9.5
  setenv JLAB_ROOT /site/12gev_phys
  source ${JLAB_ROOT}/ce/jlab.csh
  cd <wherever the source lives>
  mkdir build
  cd build
  cmake ..
  make


Also, some small changes to get this to compile with 4.9.5 (not affecting earlier versions):
G4MaterialPropertyVector is now a typedef of G4PhysicsOrderedFreeVector (which has existed
since 4.8.1 at least; let me know if this causes you discomfort).

This doesn't actually link yet with cmake, but it's a first step.