Added in random seed handling.
By default, G4 uses the same initial random seeds for different jobs, so that simulated results can be reproduced. If you would like to give a different initial random seed to your job, so that the results are statisticaly independent, you should specify the random seeds from a file.

The followings were added in the configuration macro to let G4RunManager to save the random seed.


#====================================
# Random seed handling
#====================================

# Inform the RunManager to systematically save the seed
/random/setSavingFlag 1
# Copy the following command line to your user macro and uncomment it 
# so as to use the seeds of the last event in last run
#/random/resetEngineFrom currentEvent.rndm
# Or use specific seeds from file 'myseed.rndm', for example 
#/random/resetEngineFrom myseed.rndm