Changed the Geant4 code so that now Q2 is calculated correctly for different incoming beam angles.
Before the incoming vector was only used to set the incoming beam, but not saved and passed to other parts of the code to be used in calculation. Now the incomming 
direction and positions that you can set with the GUI/macros is saved and passed to QweakSimEPEEvent.  Here it is used along with the outgoing particle mometum to get 
the scatted theta value (called RelativeThetaAngle) this theta is used as the theta to calculate out Q2, asymmetries, and cross section weight.