these are little scrpits take out the tracking parts of the Qweak experiment in the Geant 4 simulation.
the .in files are just like the ones without the _noDC, but they use the 
noTracking.mac to move the drift chambers out of the way.

macros/noTracking.mac - sets the positions of the HDC, VDC, and Trigger 
scintillator so that they are out of the way of the electrons.