Rolling back revision 5348 after explanation from Peiqing about the two different QE schemes.
Quote from Peiqing:

"1. Michael wrote the initial PMT quantum efficiency code for the main detectors. The method he used is to 
count photons which cross the physical volume of PMT window and reached the photocathode volume, and then weight the 
passing photons with the quantum efficiency of PMT photocathode according to photons' wavelength. The code was written 
via the UserInformation class.

2. I kept Michael's code, but added in an additional new method in the UserSteppingAction class. This method utilized 
the build-in photoelectric effects functions of photocathode surface in Geant4, and included all of the photocathode 
surface properties. All photons were treated according to the PMT's QE and photon's wavelength/energy."