Sometimes, you start with setting up MD9, but then you step back and decide to turn MD9 off again.  Oops, you have lost access to the commands that affect ALL main detectors...  You are confused (it worked earlier in your macro, after all).
The CerenkovDetector messenger commands that are applicable to ALL main
detectors (changing of materials of pre-radiators etc) are now static
members of the messenger.  Since they are singletons they should really
be using auto_ptr, but I'll leave fixing that memory leak as an exercise
to the reader (or the operating system, whoever comes first).

This fixes the following observed issues:
- /Cerenkov/SetContainerMaterial will work again (*)
- calling the macro twice in a row will produce the same results

(*) SetContainerMaterial never actually did anything, even though it was
a valid command.  I added the necessary accessors in the CerenkovDetector
class.