Fixed the rate and yield variables in the Cerenkov.PMT branch of the rootfiles.  They were all low by a factor of 2, because they are normalized to the number of events thrown.  They weren't taking into account the fact that, in the simulation, two particles are created for each event, one before the vertex and one after the vertex.  This affects all rate and yield variables and all reaction types.
Also added a new lookup table for 1160 MeV on LH2.  This new lookup table is valid up to 15 MeV away from the elastic peak, and will return a zero cross section past that point.  This lookup table covers from 15 MeV to 1160 MeV in E'.

Added a new cross section variable and corresponding rate and yield variables that represent the elastic peak.  In the band of phase space between the elastic peak and 15 MeV below the elastic peak, this new cross section returns the elastic peak cross section divided by 15 MeV.  This normalization is done to keep it in the same units as the lookup table cross sections.  Outside of this range, the cross section is set to zero.  This new process should be added together with all of the other processes (elastic, inelastic, quasi-elastic, pion, & US & DS windows) to get the total rate.

The new cross section is in the primary branch with all the other cross sections and is named CrossSectionRadElasticPEak.  The new Rate and yield variables are named with a similar convention (PMTTotalRateELPeak).