Write the random state of the simulation engine for each event into the ROOT tree.  This should allow to start the simulation at a particular event for GUI inspection.
Motivation: in some cases (the Moller generator) there is a very low efficiency of successful
events, and wouldn't it be nice if we didn't have to click through 10000 events in the viewer
to see what's going on with that one event that made it?


With the built-in command
  /random/setSavingFlag true
you can write the random state for each event to a file, hard-coded as currentEvent.rndm.  If the
simulation crashes, you can restore the status with
  /random/resetEngineState currentEvent.rndm
This works well, but only if the simulation crashes.  If it doesn't crash, the next event will
just overwrite currentEvent.rndm with its own random state.  Also, the frequent file writes are
horribly inefficient on the farm (slowing down jobs by factors of two).

The file currentEvent.rndm contains a vector of 202 long ints.

An attempt was made to use the built-in CLHEP::HepRandom::getTheSeed() function, which returns a
single integer, but it appears that this seed is really a constant.  It is the seed, but it isn't
updated for every event.  This, therefore, appears useless.  You can use the seed to replay an
entire run, but not an individual event in that run.  This is the approach that remoll for MOLLER
uses, but it doesn't do what we want.

Back to setSavingFlag output then.  There is an obscure function of G4RunManager that stores the
content of currentEvent.rndm to a string in G4Event: StoreRandomNumberStatusToG4Event(int).
Depending on the argument it stores the random state (1) right before calling GeneratePrimaries,
(2) right after exiting GeneratePrimaries and starting tracking, (3) in both cases.  Loading the
random state puts us right before GeneratePrimaries, so that's where we want to save the state
as well (i.e. int = 1).

From G4Event we make sure that the random seed gets to the ROOT tree in Primary.RandomSeed.

ROOT is not as good at plotting multi-line strings, or doing anything useful with them, so a
small helper function is provided that writes the file currentEvent.rndm when you put it in a
Draw or Scan command: WriteRandomSeed().  It returns the size of the file, but that's irrelevant.
You could use this as follows:
  QweakSimG4_Tree->Draw("QweakSimUserMainEvent.Primary.WriteRandomSeed()",
    "Cerenkov.PMT.PMTTotalNbOfPEs>50")
to write out currentEvent.rndm for all events with more than 50 PEs.  To avoid overwriting all these
files, they are written to event####.rndm with #### equal to the PrimaryEventNumber.  If you run
this for 10000 events, you WILL get 10000 currentEvent.rndm files in your directory :-)

For some bizarre reason you do need to specify the full branch info for the function, even though
for other leaves you can just use the short version (e.g. CrossSection, instead of the longer
QweakSimUserMainEvent.Primary.CrossSection).

Now, for those who are still reading (Jim), this unfortunately still doesn't do what we want.  If
we create start from one of the event####.rndm files that were written out, we don't get that event
back.  Somewhere there is something missing.  I can't figure it out, but I'll throw this out there
for others to explore as well.