IMPORTANT: This commit affects what's in the tree.
With this commit the prescattering energy loss events are NOT written to the ROOT file anymore.
Maybe that's what you have always wanted, but if this is a big problem for you, please let me
know.


This also fixes the random state problem.  The random state is only written for the prescattering
energy loss events, NOT for the actual events.  This ensure that when you use this state to throw
an event, it will throw the prescattering energy loss event with the appropriate random state.

1. Run a simulation in batch mode as you usually would.
2. In QweakSimRoot, run QweakSimG4_Tree->Draw("QweakSimUserMainEvent.Primary.WriteRandomSeed()")
   with whatever selection you wish to apply.
3. Rerun the simulation interactively, set up the exact same geometry (!), and run the command
     /random/resetEngineFrom event####.rndm
   with #### an event number you are interested in.