Some updates to the dummy target implementation:
1) Change the setting of the window length to the stored myUserInfo->TargetThickness instead of hard coded numbers.

2) Change the settings of the SetTargetLength, which control where the dummy targets are positioned, such that the US face of the targets are aligned (as indicated in Target ELOG 18).  For simplicity, I align the US face of the upstream targets at the front of the cell material (-0.5*targetCellInnerLength), and the US face of the downstream targets at +0.5*targetCellInnerLength.  Note that technically this is not 100% correct - see ELOG 18: "The upstream face of the upstream targets were on average 1.39 mm upstream of the LH2 entrance window" and "the downstream targets were 2.97 mm upstream of the LH2 cell exit window."  But I think it's close enough.

One last comment for the dummy targets.  The value for SetTargetLength is now calculated using targetCellInnerLength.  In the SetTargetLength function, the targetCellInnerLength is updated to this new value.  Therefore, if you call the creation of a different dummy target after the initial dummy target is created, the calculation of the position will be wrong.  So if you need to run multiple dummy targets you must run each from a fresh start.