Added a second tracking package to the simulation. The commands which were already present will only affect tracking package one. For example,
/VDC/SetFrontCenterPositionInY

will only move tracking package 1. I have added in commands for individual package control. Such as,

/VDC/VDC1/SetFrontCenterPositionInY
/VDC/VDC2/SetFrontCenterPositionInY

which will set the y positions of the VDCs for package 1 and 2 respectively. All position sets and phi rotations have individual package versions.

The only previously present commands which I changed are
/HDC/SetRotationAngleInPhi
/VDC/SetRotationAngleInPhi

which will set the phi angles for both packages. Package 1 for the angle specified, and Package 2 180 degrees for the opposite octant.

By default the second package is disabled, with both the HDCs and VDCs moved out of the way. This is done in myQweakConfiguration.mac
New fields in simulated root file to keep information on which tracking package was hit. The qwsimtracking has not been updated yet.
Currently qwsimtracking assumes all hits in package one, but will within the next couple of weeks be updated.