Modified the target class such that the scattering chamber window is now a sensitive detector and hits in it are recorded in the ROOT tree.
To compile you will need to rerun cmake:
  cd build
  cmake ..
  make

The ROOT file has an extra branch Target.Detector with such goodness as
  Target.Detector.ParticleType
  Target.Detector.GlobalPhiAngle
  Target.Detector.GlobalThetaAngle
  Target.Detector.LocalVertexTotalEnergy
  Target.Detector.LocalVertexKineticEnergy

Target energy loss (-4.56 MeV in a small simulation):
  QweakSimG4_Tree->Draw("Target.Detector.LocalVertexTotalEnergy-Primary.OriginVertexTotalEnergy",
    "Target.Detector.ParticleType==11 && Target.Detector.NbOfHits==1")

Theta angle change in target (zero with RMS of 0.09 degrees in a small simulation):
  QweakSimG4_Tree->Draw("Target.Detector.GlobalThetaAngle-Primary.OriginVertexThetaAngle",
    "Target.Detector.ParticleType==11 && Target.Detector.NbOfHits==1")

For simplicity, both of these last commands are restricted to 1 hit per event. The additional
hits per event are photons (22), although there are events with 2 or 3 elec/positrons as well.

Geant4 returns the particle type as an unsigned value and an additional function call is
required to get to the charge, so there is no way to distinguish electrons and positrons
right now.  I'll let whoever is interest in this figure out the best way to encode this.
The same is true for other detectors as well.