An update to my last commit for the Trigger Scintillators. Added in control for the second Trigger Scintillator. Can set position using commands such as
/TriggerScintillator/TriggerScintillator2/SetCenterPositionInZ 200 cm

By default, /TriggerScintillator/SetCenterPositionInZ will only set the position of the first trigger scintillator. The coordinate passed to the command should be in local coordinates, or the coordinates of the TS when in the top octant. Also fixed an issue with the trigger scintillator rotations.

Updated myQweakConfiguration.mac with only a couple of comments.

Updated macros/noTracking.mac, which now moves both trigger scintillators out of the way along with the HDCs and VDCs.