Sometimes (i.e. very rarely, but it's possible) the elastic proton generator will change the scattering angle to the minimum value, in particular when the scattering angle is below that minimum value.  We should be notified of this.
When does this have a chance of happening?
- the beam is given a non-zero angle
- there is angle straggling on the way to the primary vertex
- the scattered electron polar angle is randomly picked between min and max theta
- the scattering angle is calculated between the straggled and scattered electron
AND this scattering angle is smaller than the minimum scattering angle.

The minimum scattering angle is set by default to 4 degrees.  In a simulation of
7000 events this affected only one event.

Why am I writing all this?  I should just get rid of that arbitrary cut-off and
make it a hard 0.01 degree cut-off like for the other generators...  The reason
is just to introduce a regulator.