We shall prevail over bugs and inconsistencies!
I think this will finally generate results that make more sense to Hend and Steve.

It helps to not mix degrees and radians (to my defense the code is converting back and
forth in various levels of successive function calls).  Here I finally think the elastic
scattered energy is calculated correctly using the angles in degrees.

While I was debugging this, I also fixed the issue with using scattering angle versus
scattered electron angle.  The latter ignores the angle straggling in the hydrogen AND
the explicitly set angle by the user (that's the one that's relevant here, even if not
used currently).  Now the scattering angle is calculated properly as the angle between
the beam and the scattered electron in the scattering chamber exit window.

I'll upload a plot of the energy loss to https://qweak.jlab.org/elog/Ancillary/148.
This is to be compared to attachment 7 in https://qweak.jlab.org/elog/Ancillary/147.