Work on getting QwGeant4 to compile with geant4.10 (not complete yet).
The main (annoying) difference between geant4.9 and geant4.10 is that the system of units
and the physical constants are not loaded by default.  This means that in many places this
needs to be added by hand.  It also means that innocuous variables like 'm' suddenly start
generating shadowing errors.  Here are two include files that load the relevant units and
avoid most shadowing errors.  Next commit will be to fix the variables in the code and
update the headers.