Work on getting QwGeant4 to compile with geant4.10 (not complete yet).
The stuff that is needed to get this to compile in 4.10.  That's the most accurate
description I can come up with.  Removing variable name overlaps: m*m is surely
something you can call mm, right?  Wrong, that's a unit of distance.  Also, A, Z, M
are defined as class variables but don't need to be.