Another interesting PMT left/right difference.
There are bevels (a.k.a. chamfers) on the 4 longest edges of the quartz bars in
the geant4 simulation. According to specs [1] these chamfers should be at most
0.5 mm. Nevertheless, the simulation has them at 1 mm. Ok, no big deal so far,
and not something I'm going to change without talking to Dave Mack.

However, there is also a small angle on the outside upstream chamfer (only),
for whatever reason, that nearly makes the chamfer disappear on one end of the
half bar and makes it larger on the other end of the half bar. This angle is
hardcoded as 0.021486 degrees. Again, no big deal, probably this helps avoiding
photons bouncing around forever (incidentally, the upstream face is coded to be
1 mm taller than the nominal 18 cm, and the downstream face is 1 mm shorter,
which is probably done for similar reasons).

The problem comes in with how the small angle 0.021486 degrees is applied for
the left and the right part of the bar: they are angled in the same direction!
The light therefore propagates in the direction of reducing chamfer for one
half of the bar, while it propagates in the direction of increasing chamfer
for the other half of the bar. I've inverted the sign for the left half of the
bar, so now the slant in the chamfer is symmetric between left and right.

[1] https://qweak.jlab.org/doc-private/ShowDocument?docid=228